backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videoagent"
version = "0.1.0"
description = "Backend library for LLM-powered personalized video generation."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "VideoAgent"}]
dependencies = [
  "google-genai>=1.0.0",
  "google-cloud-storage>=2.14.0",
  "pydantic>=2.0.0",
  "moviepy>=1.0.3",
  "python-dotenv>=1.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "openai-agents[litellm]==0.8.0",
  "streamlit>=1.37.0",
  "streamlit-javascript>=0.1.5",
  "requests>=2.32.0",
  "tqdm>=4.66.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
